refactor: add clean architecture folders
