+++ src/app/page.tsx
@@ -2,16 +2,16 @@
 
 import { useState } from 'react'
 import { Bars3Icon } from '@heroicons/react/24/outline'
-import Sidebar from '@/components/Sidebar'
-import Dashboard from '@/components/Dashboard'
-import Chat from '@/components/Chat'
-import Tasks from '@/components/Tasks'
-import Notes from '@/components/Notes'
-import Goals from '@/components/Goals'
-import Calendar from '@/components/Calendar'
-import Events from '@/components/Events'
-import Settings from '@/components/Settings'
-import Profile from '@/components/Profile'
+import Sidebar from '@/presentation/components/Sidebar'
+import Dashboard from '@/presentation/components/Dashboard'
+import Chat from '@/presentation/components/Chat'
+import Tasks from '@/presentation/components/Tasks'
+import Notes from '@/presentation/components/Notes'
+import Goals from '@/presentation/components/Goals'
+import Calendar from '@/presentation/components/Calendar'
+import Events from '@/presentation/components/Events'
+import Settings from '@/presentation/components/Settings'
+import Profile from '@/presentation/components/Profile'
 
 type ActiveView =
   | 'dashboard'
