+++ src/infrastructure/agent-functions.ts
