+++ src/presentation/components/FreechAvatar.tsx
